+++ esp-backtrace/build.rs
@@ -1,6 +1,17 @@
 use esp_config::generate_config_from_yaml_definition;
 
 #[macro_export]
+macro_rules! assert_unique_features {
+    ($($feature:literal),+ $(,)?) => {
+        assert!(
+            (0 $(+ cfg!(feature = $feature) as usize)+ ) <= 1,
+            "At most one of the following features must be enabled: {}",
+            [$($feature),+].join(", ")
+        );
+    };
+}
+
+#[macro_export]
 macro_rules! assert_unique_used_features {
     ($($feature:literal),+ $(,)?) => {
         assert!(
@@ -15,8 +26,17 @@ fn main() {
     // Ensure that exactly a backend is selected:
     assert_unique_used_features!("defmt", "println");
 
-    if cfg!(feature = "custom-halt") && cfg!(feature = "halt-cores") {
-        panic!("Only one of `custom-halt` and `halt-cores` can be enabled");
+    // Ensure that there aren't multiple halt methods selected:
+    assert_unique_features!("custom-halt", "halt-cores", "semihosting");
+
+    if !cfg!(feature = "panic-handler")
+        && cfg!(any(
+            feature = "custom-halt",
+            feature = "halt-cores",
+            feature = "semihosting"
+        ))
+    {
+        print_warning("A halt method is selected, but esp-backtrace is not the panic handler.")
     }
 
     // emit config
@@ -24,4 +44,8 @@ fn main() {
     let cfg_yaml = std::fs::read_to_string("./esp_config.yml")
         .expect("Failed to read esp_config.yml for esp-backtrace");
     generate_config_from_yaml_definition(&cfg_yaml, true, true, None).unwrap();
+}
+
+fn print_warning(message: impl core::fmt::Display) {
+    println!("cargo:warning={message}");
 }