+++ esp-hal/src/uart/mod.rs
@@ -42,9 +42,8 @@
 //! [embedded-io-async]: embedded_io_async
 
 /// UHCI wrapper around UART
-// TODO debug C3/S3 to remove the device cfgs
 // TODO add support for PDMA and multiple UHCI for 32/S2 support
-#[cfg(all(soc_has_uhci0, gdma, any(esp32h2, esp32c6)))]
+#[cfg(all(soc_has_uhci0, gdma))]
 #[cfg(feature = "unstable")]
 pub mod uhci;